fix(cli): get run mode working via CLI
