+++ lib/index.cjs
@@ -11,10 +11,7 @@ var semver = require("semver");
 var platformOverrides = require("./platformOverrides.cjs");
 var deprecate = require("deprecate");
 
-const {
-  checkCache,
-  Platforms,
-} = require("../dist/index.cjs");
+const { checkCache, Platforms } = require("../dist/index.cjs");
 var NwVersions = require("./versions.cjs");
 var Version = require("./Version.cjs");
 var Utils = require("./utils.cjs");
@@ -24,7 +21,6 @@ var Downloader = require("./downloader.cjs");
 inherits(NwBuilder, EventEmitter);
 
 function NwBuilder(options) {
-
   this.options = options;
   this._platforms = _.cloneDeep(Platforms);
 }
