+++ src/index.js
@@ -5,11 +5,4 @@ import Platforms from "./constants/Platforms.js";
 import checkCache from "./utilities/checkCache.js";
 import parseOptions from "./utilities/parseOptions.js";
 
-export {
-  nwbuild,
-  Options,
-  Platform,
-  Platforms,
-  checkCache,
-  parseOptions,
-};
+export { nwbuild, Options, Platform, Platforms, checkCache, parseOptions };
